generate: Do not touch files with checksums
Overwriting checksums is not very useful, and can accidentally cause us
to lose valuable checksum information.

Change the generate subcommand to only compute checksums for files that
don't have one.

This is very useful when generating checksums for a lot of files for the
first time, because it will not waste time on files that already have
been processed.