test: Add mini-sleeps before changing files
On fast systems, the resolution of the filesystem mtimes may not be high
enough to distinguish between two file editions, even with summer runs
in between.

This causes tests to be flaky with false positives.

To fix this, we introduce very brief sleeps between changing files, just
long enough to ensure the mtime will have changed (on a reasonably modern
filesystem, under normal conditions).