Extract walking logic to a common function
All the main operations (update, generate, verify) have the same
structure: a per-file function, and then surrounding logic to walk the
roots.

Today, the latter is duplicated on each function.

This patch abstracts it away, so each per-file function is more self
contained, and making it easier to make changes to the walking logic
(which will happen in later patches).