Parallelize walking logic
This patch makes walk() run in parallel by default.

Depending on the media and the nature of the files, this can make things
faster or slower.

Simple, unscientific experiments comparing serial to parallel show large
improvements for data in cache or on SSD, and mixed results on HDD
(depending a lot on the drive and the data itself):

       | before | after | difference
cached |   8.8s |  5.2s |       -40%
ssd    |  28.4s | 12.6s |       -56%
hdd 1  |  12.1s | 14.8s |       +22%
hdd 2  |  23.2s | 18.4s |       -20%