github: Disable cache in govulncheck, always use latest version
The govulncheck action is supposed to use the latest Go release, but often
it does not due to an older one being cached in the runner.

This causes false positives when there are security issues in the Go
distribution itself.

To help reduce those, disable the cache and specify that we want the
latest stable version. We need the latter to nudge the setup-go action to
actually fetch the latest one.