Implement support for working on a random subset of files
This patch implements a new `-subsetpct` flag, that makes summer only
work on a randomly selected subset of the files, targeting to include
the given % of files.

Due to randomness, the actual files selected will be different each run,
and can be higher or lower than the given target percentage.

This is mainly useful to verify a random selection of files (for
example, to schedule a verification of a random 20% of files per day).