ui: Add workaround to prevent display of repeated characters
When outputting to a TTY (where we do \r), today there are situations
where the string representation of a duration shortens, and that can end
up with us repeating one or two characters at the end of the line.

This can happen in the following situations:

1m59s: 0 matched, 0 modified, 0 new, 0 corrupted
2m0s: 0 matched, 0 modified, 0 new, 0 corruptedd     <-- extra 'd'
1h59m59s: 0 matched, 0 modified, 0 new, 0 corrupted
2h0m0s: 0 matched, 0 modified, 0 new, 0 corrupteded  <-- extra 'ed'

In this patch, we make summer add two spaces at the end of each line, as
a workaround for when the duration is shorter than the previous one, the
old characters are still visible.

For verbose output, or non-TTY, this is not needed as we never overwrite
lines anyway.