Print checksum mtime on corrupted files
When a file is corrupted, it can be useful to know the checksum mtime
for debugging purposes.

Today, summer doesn't show it, so this patch adds it to the information
printed when a file is corrupted.

It now prints the epoch as well as the parsed time, for extra
convenience.